Only honor GPS overrides INSIDE the perimeter.
On a large lawn, furthest from the perimeter-sender,
after crossing the perimiter wire, the magnitude might
quickly drop-off, preventing us from honoring
the 'outside-ness' of the signal.

This prevents the GPS flagInsidePolygon from being
activated if the current perimeter-sensor magnitude
indicates that we were/are outside the perimeter.

This is especially useful because GPS coordinates can be
up 6+ meters inaccurate, making a GPS bounding box intersect
with boundary wire sections.